pyproject.toml
[project]
name = "ecomify"
version = "0.1.0"
description = "A modern e-commerce backend featuring a robust RESTful API powered by Django and DRF."
authors = [
    {name = "Cesar Gonzalez",email = "cesargonzalez390@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
django = "^6.0"
black = "^25.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
